firmware_v2: configure clock tree and blink LED
Set up the clock for the HSE external xtal (8MHz) feeding
the PLL with 9x multiplier yielding 72 MHz SYSCLK.

I *think* the default was for HSI at 8MHz with a slightly
lower clock frequency (48 MHz?) but it's a bit hard to tell.
The default is set via SystemInit() which is called before main().

SystemCoreClockUpdate() must be called before HAL_Init() so
that the SysTick timer is set up properly.  If this is not done,
our 200ms blink looks more like 1s.