firmware_v2/console: use interrupt-driven tx
Establish a 1K circular transmit buffer and manage it
with interrupts.

Again, bypass the HAL interrupt code and provide our own.
The HAL code assumes a fixed buffer length will be
transferred, but we want a circular buffer.

If the buffer is full, characters are dropped.
TODO: report that to the caller, ultimately as a CAN NAK.