hardware: update eurocard schematic with flip flop
Add 74AUP1G74 D flipflop with preset and clear in front
of high side load switch.  The goal is to isolate the
load switch from a reset on the STM32, for example if
it locks up and has to be reset, or if it requires a
firmware update.

Set up CLR with its own standalone power-on RC circuit so
that it does not power on the load when power is first applied
Note that the AUP series version of this chip has schmitt
trigger inputs, thus is appropriate for the slow edge of the RC.

Toggle the state of the load switch with rising edge of CLK.

Sense the state of the load swtich with D output.
N.B. The INA169 could stand in for this but not if the pi
is not plugged in drawing current.