hardware: add some power control discretes
Add 10K pull-down on PWR_TOGGLE input, so that it
doesn't float high and inadvertently trigger the flip-flop
when the STM32 puts GPIO in high-Z during reset.

Change power-on-reset cap from 100nF to 1uF.  Anecdotally,
a longer time constant is needed to wait for stable inputs.

Add R_l = 10K to INA169 for gain of 10X.

Fix a value/label error on MAX7219 10uF cap.