hardware: add jumper to force pi power on
Add J1 jumper to set 74AUP74 ~PRE input L, thus forcing
the pi power to remain on regardless of PWR_TOGGLE input.
This may be useful for using a pi carrier as a slot 0
controller, where the pi should not be under power
management with the rest of the cluster.