hardware/bus-3: add first bus prototype
This 3-slot version of the molliebus backplane was sent
to oshpark for fab.

GA[0:1] are hardwired no the board and GA[2:5] are jumperable.

5V power is supplied directly by a Molex junior fit connector.
There is no on-board fuse.

The C row is brought out to 2.54mm headers for each slot.

Jumper selectable CAN termination is available at both ends of
the bus.

Unfortunately not much thought was given to chaining busses so
~RESET and CAN+/- were not brought out to connectors.