hardware/dodec_bracket: bracket to attach crates
This bracket allows multiple salvaged and modified Elma 3U
eurocard crates to be attached together in a dodecagon
configuration.  The racks are oriented vertically (sideways)
with fronts facing outward and backplanes clustered inward.

The end result is roughly "cray shaped".  The Cray-1 was a
hexadacagon with 12 of the 16 slots populated.  This dodecagon
with 9 of the 12 slots populated is pretty close.

Using 8 of the 9 crates for compute reaches 128 nodes at
16 nodes per crate.

The machine fits nicely on a 31.5 inch diameter round coffee table.