hardware/40pin-adapter: reduce footprint, add LED,term
The pcb is larger than it needs to be.  Reduce the footprint
to just outside of the header pads, and move vias inward.

Also, may as well add a power LED here where it's more visible
than on the main board next to the connector.

I2C is perhaps better terminated here than the main board too,
since it's closer to the end of the bus, so add two 4.7K resistors.

We can always skip populating if the main board works out better.