firmware/matrix: expose matrix/led interfaces
Allow the display(s) to be updated by other parts of the firmware
asynchronously.  Setter functions atomically update a frame buffer,
then notify a task to clock the frame buffer out to the displays
over SPI.

If multiple updates occur before the task can update the display,
the last in wins.

Move the "address display" function over to an init task in main().
Have that task access the address and use the above interface to display
the two digits with delays.

Add a function + task for pulsing the green LED (e.g. for CAN activity).
Add a function for setting the value of the red LED.