firmware/power: don't change pi power on restart
Problem: resetting the STM32 causes pi power to be lost.
This is not consistent wtih the hardware design.

Don't reset the power state.

N.B. There may be a problem with the power on reset logic that
results in pi being initially on that we'll have to deal with
(which is what prompted setting the initial state to off earlier).