pi-carrier: better proto area
Modify protopad footprint for 8x8 2.54mm grid, plus two 1x8 rows
of connections.  Maps to a 2x8 connector.

Add a row of grounds on the left.
Make N.C. pin on the right a ground.
Add silk labels.
Reroute to accomodate larger area.

Route 5V (unswitched) to be slightly more direct to the
LED Matrix subsystem than it was, and use back of board for vert.
like others.