pi-carrier: replace MCP2515 XTAL with CMOS osc.
Problem: XTAL is rather large and a through-hole component.

Substitute Abracon ASE-8.0000MHZ-LR CMOS XTAL oscillator in
a 1.2mm x 2.5mm SMT package.  The LR is
 L: -40C to +85C (widest temp range)
 R: +/-40ppm (2nd best stability)

Add grounded DNP resistor footprint on the OSC2 input of MCP2515 just
in case, since datasheet suggests this as an option to reduce noise.

This oscillator is not shared with the one on the STM32 because
it is powered by PI_3V3.  It didn't seem like a good idea to mix
power domains.