40pin-adapter: drop LED and term, update footprint
Problem: The LED and i2c terminators need to be on the
main board, not the adapter, leaving open the possibility of
using an IDC ribbon cable instead.

Drop them from the adapter and update the version to v4.

Adjust schematic to letter size and fill in title, etc..
Update PCB.