firmware/serial: add usart1 support
Add support for reading/writing usart1 which will be attached
to the pi console.

Upon rx interrupt, data is stored in a stream buffer.
The stream buffer "triggers" on 8 characters, the maximum
CAN payload.

serial_recv() reads from the buffer with timeout.

serial_send() sends one or more characters directly to the
serial port, blocking while the peripheral is busy.