firmware/canbus: add console support
Add support for pi console.

Implement the canmsg_v1 console protocol.

Upon a receiving the CONSOLECONN (connect) request, start a
task that reads from the serial port and writes to the
registered object (normally CONSOLESEND).

The task is suspended after every DAT send, and resumed upon
receipt of an ACK, thus enforcing lock-step of DAT/ACK as a form
of flow control.