firmware: make serial overruns more "chunky"
Problem: once the serial console buffer fills up, you get
nothing but unreadable garbage until the receiver catches up
a bit.

Stop storing characters in the buffer until the receiver
has freed up a threshold number of bytes (256) in the buffer.
The net result is that a chunk of output goes missing, not
every other character.