drivers: add CAN device tree overlay
Import mcp2515-can0-overlay.dts from raspberry pi linux-5.10
source tree and modify for SPI bus 1, device 0.

Enable with the following in config.txt:

  dtparam=spi=on
  dtoverlay=spi1-1cs
  dtoverlay=mcp2515-can2,interrupt=12

This assumes the MCP2515 on the pi-carrier board is running with
a 16 MHz clock, necessary to achieve desired 1m baud CAN rate.
(Initially an 8 MHz clock was specified).

The can0 is brought online with:

  ip link set can0 up type can bitrate 1000000