bramble: install CAN filters
Problem: the raw CAN interface is configured without filters,
causing unnecessary wake-ups for CAN processes such as the power
and console helpers.

Add an argument to can_open() to allow the local address to be
specified, and then install filters for that destination address
and the broadcast address.

Add can_open_with() which accepts an array of filters.  The filter
array is empty, the interface is effectively in promiscuous mode
like before.

Update utilities.