CAN protocol: add eot flag
Problem: A sender of streaming data (for example, a control processor
sending console data) needs a method to signify that a message is the
last in a batch so that the receive end can send an ACK immediately.

Reduce the range of 'seq' to 16, and add an 'eot' flag to the
message id.

Update CAN.md protocol doc, and encode/decode functions in bramble
and firmware.