bramble/console: reorder DAT messages on receive
Problem: occasionally we observe messages arriving out
of order.

Receive an entire sequence, to final message with EOT bit
set, before relaying to standard output.  If messages in
the sequence arrive out of order, place them in order.

Set a timer on receipt of seq=0.  If it expires before
the sequence is complete, then:
- if the final eot was not received, abort and let
  conman restart the connection
- if another message was not received, print what we have
  and continue on