firmware/power: use pi GLOBAL_EN, RUN_PG
Problem: software power off, where the pi turns off its own PMIC,
is not reflected in CAN power state or red LED.

A better way to control power than externally controlling the 5V input
is to use the pi 4 J2 RUN_PG and GLOBAL_EN pins.  This requires fewer
components for future board versions, and power status can reflect the
state of the pi PMIC rather than the external power.

The following fun facts about RUN_PG and GLOBAL_EN are pertinent:
- RUN_PG is at 3V3 when the pi PMIC has completed its power on sequence
- RUN_PG is at GND when the pi PMIC is off
- RUN_PG may be pulled to GND momentarily to reset the pi without shutting
  down the PMIC.
- GLOBAL_EN is pulled to 5V in the pi, so a 5V tolerant pin is required
- Setting GLOBAL_EN low turns off the PMIC
- Letting GLOBAL_EN be pulled high starts the PMIC power on sequence
- If the PMIC is turned off by the pi, GLOBAL_EN can be pulled low for >= 1ms
  to restart the PMIC.

This firmware change assumes the following hardware changes to v1 and v2
pi-carrier boards:
- Connect RUN_PG to PA8
- Connect GLOBAL_EN to PB14
- disconnect PI_5V from power switch output
- connect PI_5V to +5V

The flip-flop still provides the initial power state when STM32 comes out of
reset so that resetting the STM32 doesn't affect the pi if it happens to be
running.

The REMOTE/ON jumper forces the flip flop to the ON state.  Consequently,
if set to ON, GLOBAL_EN is set to 1 on power up or STM32 reset.  For the
most part this has the desired effect, except:
- If pi PMIC is turned off via CAN, it is turned back on upon STM32 reset
- If pi PMIC is turned off via pi, it is not turned back on upon STM32 reset

Remove commented-out power_set_state () from firmware main ().

Note that power_set_state() can now put the calling task to sleep.
It is called from the CAN receive thread, so there is a potential for
overruns there.  Right now I don't think that is a problem.