bramble: add temperature to bargraph
Problem: under load, the ARM core temperature becomes
an interesting parameter to visualize, and we have one
extra column on the 5x7 matrix after displaying utilization
of four CPUs.

Make the 5th column represent temperature in a range of
30C - 85C.  The pi starts throttling clock speed at 85C,
so that is effectively a maximum, and 30C is a fairly reasonable
idle temperature.

Add a handler for SIGTERM that clears the display so when
the pi is shut down it doesn't leave a residual image.