firmware: add support for SHUTDOWN
Problem: there is no way for the service processor to nicely shut
down the pi.

Implement the SHUTDOWN CAN object.  This utilizes two GPIO lines
on the Pi:

Pi GPIO27 (STM32 A4) is configured so that when it transitions low,
systemd captures a poweroff keypress.  Add to config.txt:

  dtoverlay=gpio-shutdown,gpio_pin=27,active_low=1,gpio_pull=up

Pi GPIO26 (STM32 A5) is configured so that when the OS is running,
it is driven high as an output, and when the OS is stopped or power
is off, it goes to its natural state (pulled low).  Add to config.txt:

  gpio=26=op,dh