pi-carrier: eliminate BOOT1 jumper
Problem: BOOT[0:1] jumpers are not needed because we program the
chip with SWD, either from the SWD header with an external stlink-v2
or from openocd on the pi.

Boot modes are as follows

BOOT0 BOOT1 Mode
0     0     User flash (default mode)
1     0     System memory (bootloader)
1     1     SRAM memory

Tie BOOT0 and BOOT1 to ground through a 100K resistor as on the
blue pill.

If we need to bodge something later to select a different mode,
the pads for the series resistors can be used for that.