hardware: change kludge panel holes for Elma
Problem: kludge front panel starting point is for the Reichelt
BGT VS1, but we have Elma S-216-13 eurocard hardware on hand.

Adjust position of two mounting holes in vertical direction so
that they are 13.25mm from the top and bottom edge.  Adjust the
horizontal position of one hole so that it matches the lever
mounting bracket.  Replace the handle hole with a slot for the
lever, which also serves as a handle.

Goal will be to replace Elma hardware with cheap DIY 3D printed
alternative but use the same hole pattern.

Add a README referencing Marten's work.