hardware: add pi front panel
Problem: pi carrier 3d printed PLA front panel is not sufficiently
rigid and does not support an ejector lever.

Have a go at converting the kludge card PCB front panel for use
with the pi-carrier.  The pi holes are in a separate footprint,
since it was easier to translate the FreeCAD parameters to a zero
origin footprint.