firmware: save power state in backup register
Problem: GLOBAL_EN glitches on a STM32 warm reset, sometimes
turning on the pi PMIC.

On each power change, save the new power state in a backup data
register that is preserved across a warm reset.  Initialize
GLOBAL_EN to that value if not coming out of a power-on-reset.

Fixes #64