build: build firmware with autotools
Problem: To make this more like one project, the firmware should
be built with autotools.

The firmware will be configured as an autoconf subproject with
a cross toolchain.  Pull in the AX_SUBDIRS_CONFIGURE macro from the
autoconf-archive, and configure --host=arm-none-eabi in the firmware
subproject.

Construct a Makefile.am that uses the same build flags as before.
Note that libopencm3 continues to use its own build system, and
that there are some weak symbols in librtos that need to be overridden
by strong ones in libwwg/opencm3.c.

This hard codes the installation of firmware in /lib/bramble.fw
for now.