firmware: display single address digit
Problem: the geographic slot address is displayed as two
digits, one after another, which is not a great user interface.

Just display a single hex digit, which covers addresses up to 15.
Also, don't blank the display after showing it.  Leave it until
Linux decides to start showing the load graph.