pi5_adapter: add power switch
Problem: the Pi 5 and Pi 3B (not plus) do not expose PMIC
GLOBAL_EN line, thus cannot be controlled with the pi-carrier.
Also, on Pi4 and Pi 3B+ that do expose the PMIC, bodge wires
direct from pi to pi-carrier are inconvenient as it prevents
removal of the pi.

Add the 3A high side load switch that was eliminated from the
pi-carrier.  The power circuit emulates the pi PMIC lines on a
JST SH (1mm) 3-pin connector.  That is,

- GLOBAL_EN (pulled to 5V on the board) turns the switch on when
  floating, off when grounded.

- RUN_PG is pulled high by the pi 3V3 power output

The new components can optionally be left unpopulated and the
Pi GLOBAL_EN and RUN_PG can be soldered to pads on the baord
and brought out on the JST connector.  A bypass jumper must also
be bridged in this mode.