pi-carrier: ensure TP Reference is plotted on silk
Problem: testpoint silk labels did not show up on the v3.1
board.

I thought I was being clever and replaced the Value field of
"TestPoint" with a descriptive name, and placed it on the silk
layer, but KiCad 7 did not plot Value fields in the fabrication
outputs.

After receiving the boards, I noted that the Gerber Viewer tool
did not show those fields on the silk layer.  Maybe it is a KiCad
bug or maybe I did something wrong.

Set Value back to "TestPoint" in all TP footprints, uncheck "visible".

Place the Reference field on the silk layer, check "visible".

Now the silk labels are the reference desgnators (e.g. "TP17")
rather than the descriptive names, but this is probably more
appropriate and expected anyway.