pi-carrier: put 3V3 reg behind efuse
Problem: the efuse *might* allow for hot swapping, at least from
a current control point of view, but the linear 3V3 supply, which
is low current but has a large electrolytic capacitor, is still
directly connected to the bus.

Put it behind the efuse and relocate to the bottom of the board.