build: add hardware configure logic
Problem: kicad-cli in version 8 allows ERC/DRC and CAM generation
to be run programmatically, but our build system does not take
advantage of that.

Add configure logic in the hardare subdirectory that looks for
an appropriately recent version of kicad.

If --with-kicad is specified, not finding one is an error.
If --without-kicad is specified, kicad is ignored even if found.

Otherwise, new targets will be enabled if kicad is found, or
disabled if not.

The version check uses the AX_COMPARE_VERSION macro from the
autoconf archive.

https://www.gnu.org/software/autoconf-archive/ax_compare_version.html