pi-carrier: improve power zones
Problem: the pi 3v3 and m2 3v3 power zones are kind of
in conflict.

Reorient them so m3 3v3 doesn't require a long trace to
receive power.

Move the M.2 power supply a bit and tweak its silkscreen.

Drop the silk screen box from the M.2 socket as it is in
conflict wtih anything that overlaps it. Instead, show the
entire M.2 2280 board outline in User.notes.