pi-carrier: reroute GPIB traces
Problem: pi GPIB traces take a long route to get to the STM32
but now there is space for a more direct route.

Reroute traces.

Also put pi CAN diff pair on the back of the board.  Perhaps the
impedance will be higher since it's further away from the ground
plane.