firmware: fix pi5 power issue
Problem: after shutdown -h, the power button won't turn on a pi5.

On the pi5 we don't have access to the PMIC so RUN_PG is pulled to
the pi's 3V3 and GLOBAL_EN is connected to an external load switch.
shutdown -h turns off the PMIC and therefore RUN_PG, but GLOBAL_EN
reads high since it was never set low by pi-carrier firmware.

This is like the situation on the pi4 where GLOBAL_EN has to be pulsed
low for 1ms after a shutdown -h, but in the pi5 case we cannot signal
the PMIC directly so all we can do is remove external power to get it
back to its initial power-on state.  This takes longer than 1ms.

Raise the 1ms delay to 100ms.